requirements.txt
fastapi[standard]
sqlalchemy[asyncio]
aiosqlite
alembic
pytest-asyncio